Move tracing subscriber initializer to hermes-logger
