+++ crates/cosmos/cosmos-integration-tests/src/init.rs
@@ -7,18 +7,14 @@ use hermes_cosmos_chain_components::types::config::gas::dynamic_gas_config::Dyna
 use hermes_cosmos_chain_components::types::messages::packet::packet_filter::PacketFilterConfig;
 use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 use hermes_error::types::Error;
+use hermes_logger::subscriber::init_tracing_subscriber;
 use hermes_runtime::types::runtime::HermesRuntime;
 use hermes_test_components::setup::traits::driver::CanBuildTestDriver;
 use ibc::core::host::types::identifiers::PortId;
 use serde_json::Value as JsonValue;
 use tokio::runtime::Builder;
 use toml::Value as TomlValue;
-use tracing::level_filters::LevelFilter;
-use tracing::{info, Level, Subscriber};
-use tracing_subscriber::filter::Targets;
-use tracing_subscriber::prelude::__tracing_subscriber_SubscriberExt;
-use tracing_subscriber::util::SubscriberInitExt;
-use tracing_subscriber::{fmt, EnvFilter};
+use tracing::info;
 
 use crate::contexts::binary_channel::setup::CosmosBinaryChannelSetup;
 use crate::contexts::binary_channel::test_driver::CosmosBinaryChannelTestDriver;
@@ -44,25 +40,10 @@ impl FromStr for TestPreset {
     }
 }
 
-pub fn build_tracing_subscriber() -> impl Subscriber + Send + Sync {
-    let env_filter = EnvFilter::builder()
-        .with_default_directive(LevelFilter::INFO.into())
-        .from_env_lossy();
-
-    let target_filter = Targets::new().with_target("hermes", Level::TRACE);
-
-    tracing_subscriber::registry()
-        .with(fmt::layer())
-        .with(target_filter)
-        .with(env_filter)
-}
-
 pub fn init_test_runtime() -> HermesRuntime {
     let _ = stable_eyre::install();
 
-    let subscriber = build_tracing_subscriber();
-    // Avoid crashing if already initialised
-    let _ = subscriber.try_init();
+    init_tracing_subscriber();
 
     let tokio_runtime = Arc::new(Builder::new_multi_thread().enable_all().build().unwrap());
