+++ crates/logging/logger/Cargo.toml
@@ -17,3 +17,6 @@ hermes-relayer-components           = { workspace = true }
 hermes-relayer-components-extra     = { workspace = true }
 hermes-logging-components           = { workspace = true }
 hermes-tracing-logging-components   = { workspace = true }
+
+tracing                             = { workspace = true }
+tracing-subscriber                  = { workspace = true }
