+++ crates/logging/logger/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod contexts;
+pub mod subscriber;
 
 pub use contexts::{HermesLogger, UseHermesLogger};
